requirements.txt
Package               Version  
--------------------- ---------
asgiref               3.3.1    
certifi               2020.12.5
chardet               4.0.0    
Django                3.1.7    
django-cors-headers   3.7.0    
django-rest-framework 0.1.0    
djangorestframework   3.12.2   
idna                  2.10     
pip                   19.2.3   
pytz                  2021.1   
requests              2.25.1   
setuptools            41.2.0   
sqlparse              0.4.1    
urllib3               1.26.3   
